Chore: Moving some typings around
Fix: Changing to atomic write = no more file system race coditions
