+++ src/typings/atomicWrite.ts
@@ -1,0 +1,6 @@
+interface AtomicWriteOptions {
+  mode?: number;
+  exclusive?: boolean;
+}
+
+export { AtomicWriteOptions };
