+++ src/typings/response.ts
@@ -1,0 +1,6 @@
+interface StatusResponse {
+  ApiReachable: boolean;
+  online: { [key: string]: boolean };
+}
+
+export { StatusResponse };
